rename: delete source entry metadata only, skipping hard links

+++ weed/mount/meta_cache/meta_cache.go
@@ -62,7 +62,7 @@ func (mc *MetaCache) doInsertEntry(ctx context.Context, entry *filer.Entry) erro
 	return mc.localStore.InsertEntry(ctx, entry)
 }
 
-func (mc *MetaCache) AtomicUpdateEntryFromFiler(ctx context.Context, oldPath util.FullPath, newEntry *filer.Entry) error {
+func (mc *MetaCache) AtomicUpdateEntryFromFiler(ctx context.Context, oldPath util.FullPath, newEntry *filer.Entry, shouldDeleteChunks bool) error {
 	//mc.Lock()
 	//defer mc.Unlock()
 
@@ -74,8 +74,14 @@ func (mc *MetaCache) AtomicUpdateEntryFromFiler(ctx context.Context, oldPath uti
 				// leave the update to the following InsertEntry operation
 			} else {
 				glog.V(3).Infof("DeleteEntry %s", oldPath)
-				if err := mc.localStore.DeleteEntry(ctx, oldPath); err != nil {
-					return err
+				if shouldDeleteChunks {
+					if err := mc.localStore.DeleteEntry(ctx, oldPath); err != nil {
+						return err
+					}
+				} else {
+					if err := mc.localStore.DeleteOneEntrySkipHardlink(ctx, oldPath); err != nil {
+						return err
+					}
 				}
 			}
 		}
@@ -110,6 +116,12 @@ func (mc *MetaCache) FindEntry(ctx context.Context, fp util.FullPath) (entry *fi
 	}
 	mc.mapIdFromFilerToLocal(entry)
 	return
+}
+
+func (mc *MetaCache) DeleteEntrySkipHardlink(ctx context.Context, fp util.FullPath) (err error) {
+	//mc.Lock()
+	//defer mc.Unlock()
+	return mc.localStore.DeleteOneEntrySkipHardlink(ctx, fp)
 }
 
 func (mc *MetaCache) DeleteEntry(ctx context.Context, fp util.FullPath) (err error) {
